requirements.txt
dashscope==1.14.1
faiss-cpu
gradio==4.19.2
langchain==0.1.10
langchain_community==0.0.25
Pillow==10.2.0
python-dotenv
sentence-transformers
torch==2.2.0
transformers==4.36.0